algolia/bin/requirements.txt
algoliasearch==2.1.0
python-dotenv==0.11.0
